backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pydantic==2.3.0
telethon==1.29.2
sqlalchemy==2.0.20
aiosqlite==0.19.0
python-dotenv==1.0.0
